move parts of sqlite bindings to a separate module
This allows ksqlite to declare an api dependency on some parts of
sqlite bindings (e.g. ResultCode) while keeping direct sqlite
interactions internal
