+++ ksqlite3/src/commonTest/kotlin/com/birbit/sqlite3/SqliteConnectionTest.kt
@@ -15,10 +15,6 @@
  */
 package com.birbit.sqlite3
 
-import com.birbit.sqlite3.internal.AuthResult
-import com.birbit.sqlite3.internal.AuthorizationParams
-import com.birbit.sqlite3.internal.ResultCode
-import com.birbit.sqlite3.internal.SqliteException
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
