+++ sqlitebindings-api/src/commonMain/kotlin/com/birbit/sqlite3/ColumnType.kt
@@ -13,16 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package com.birbit.sqlite3
 
-pluginManagement {
-    repositories {
-        maven("https://teamcity.jetbrains.com/guestAuth/app/rest/builds/buildType:(id:Kotlin_KotlinPublic_Compiler),number:1.4.0-dev-1793,branch:(default:any)/artifacts/content/maven")
-
-        mavenCentral()
-
-        maven("https://plugins.gradle.org/m2/")
+// see https://www.sqlite.org/c3ref/c_blob.html
+inline class ColumnType(val value: Int) {
+    companion object {
+        val INTEGER = ColumnType(1)
+        val FLOAT = ColumnType(2)
+        val STRING = ColumnType(3)
+        val BLOB = ColumnType(4)
+        val NULL = ColumnType(5)
     }
 }
-include("sqlitebindings", "sqlitebindings-api", "jnigenerator", "ksqlite3")
-enableFeaturePreview("GRADLE_METADATA")
-
