+++ sqlitebindings/src/commonMain/kotlin/com/birbit/sqlite3/internal/SqliteApi.kt
@@ -15,62 +15,30 @@
  */
 package com.birbit.sqlite3.internal
 
-// see https://www.sqlite.org/c3ref/c_deny.html
-inline class AuthResult(val value: Int) {
-    companion object {
-        val OK = AuthResult(0)
-        val DENY = AuthResult(1)
-        val IGNORE = AuthResult(2)
-    }
-}
-// see https://www.sqlite.org/c3ref/c_blob.html
-inline class ColumnType(val value: Int) {
-    companion object {
-        val INTEGER = ColumnType(1)
-        val FLOAT = ColumnType(2)
-        val STRING = ColumnType(3)
-        val BLOB = ColumnType(4)
-        val NULL = ColumnType(5)
-    }
-}
-// commonized sqlite APIs to build the rest in common, or most at least
-inline class ResultCode(val value: Int) {
-    fun errorString() = SqliteApi.errorString(this)
+import com.birbit.sqlite3.Authorizer
+import com.birbit.sqlite3.ColumnType
+import com.birbit.sqlite3.ResultCode
+import com.birbit.sqlite3.SqliteException
 
-    companion object {
-        val OK = ResultCode(0) /* Successful result */
-        val ERROR = ResultCode(1) /* Generic error */
-        val INTERNAL = ResultCode(2) /* Internal logic error in SQLite */
-        val PERM = ResultCode(3) /* Access permission denied */
-        val ABORT = ResultCode(4) /* Callback routine requested an abort */
-        val BUSY = ResultCode(5) /* The database file is locked */
-        val LOCKED = ResultCode(6) /* A table in the database is locked */
-        val NOMEM = ResultCode(7) /* A malloc() failed */
-        val READONLY = ResultCode(8) /* Attempt to write a readonly database */
-        val INTERRUPT = ResultCode(9) /* Operation terminated by sqlite3_interrupt()*/
-        val IOERR = ResultCode(10) /* Some kind of disk I/O error occurred */
-        val CORRUPT = ResultCode(11) /* The database disk image is malformed */
-        val NOTFOUND = ResultCode(12) /* Unknown opcode in sqlite3_file_control() */
-        val FULL = ResultCode(13) /* Insertion failed because database is full */
-        val CANTOPEN = ResultCode(14) /* Unable to open the database file */
-        val PROTOCOL = ResultCode(15) /* Database lock protocol error */
-        val EMPTY = ResultCode(16) /* Internal use only */
-        val SCHEMA = ResultCode(17) /* The database schema changed */
-        val TOOBIG = ResultCode(18) /* String or BLOB exceeds size limit */
-        val CONSTRAINT = ResultCode(19) /* Abort due to constraint violation */
-        val MISMATCH = ResultCode(20) /* Data type mismatch */
-        val MISUSE = ResultCode(21) /* Library used incorrectly */
-        val NOLFS = ResultCode(22) /* Uses OS features not supported on host */
-        val AUTH = ResultCode(23) /* Authorization denied */
-        val FORMAT = ResultCode(24) /* Not used */
-        val RANGE = ResultCode(25) /* 2nd parameter to sqlite3_bind out of range */
-        val NOTADB = ResultCode(26) /* File opened that is not a database file */
-        val NOTICE = ResultCode(27) /* Notifications from sqlite3_log() */
-        val WARNING = ResultCode(28) /* Warnings from sqlite3_log() */
-        val ROW = ResultCode(100) /* sqlite3_step() has another row ready */
-        val DONE = ResultCode(101) /* sqlite3_step() has finished executing */
+// TODO trying to use ResultCode here crashed the compiler, hence using Ints
+internal inline fun checkResultCode(
+    dbRef: DbRef,
+    received: Int,
+    expected: Int
+) {
+    if (received != expected) {
+        throw SqliteException.buildFromConnection(dbRef, received)
     }
 }
+
+internal fun SqliteException.Companion.buildFromConnection(dbRef: DbRef, errorCode: Int?): SqliteException {
+    return SqliteException(
+        resultCode = errorCode?.let { ResultCode(it) } ?: SqliteApi.errorCode(dbRef),
+        msg = SqliteApi.errorMsg(dbRef) ?: errorCode?.let { SqliteApi.errorString(ResultCode(it)) }
+        ?: "unknown error"
+    )
+}
+fun ResultCode.errorString() = SqliteApi.errorString(this)
 
 interface ObjRef {
     fun dispose()
@@ -81,27 +49,6 @@ expect class DbRef : ObjRef
 
 expect class StmtRef : ObjRef {
     val dbRef: DbRef
-}
-
-class AuthorizationParams(
-    val actionCode: Int,
-    val param1: String?,
-    val param2: String?,
-    val param3: String?,
-    val param4: String?
-) {
-    override fun toString(): String {
-        return "AuthorizationParams(actionCode=$actionCode, param1=$param1, param2=$param2, param3=$param3, " +
-            "param4=$param4)"
-    }
-}
-
-// TODO make this a fun interface when migration to kotlin 1.4 happens
-//  maybe not as this likely needs a dispose :/
-interface Authorizer {
-    operator fun invoke(params: AuthorizationParams): AuthResult
-    fun dispose() {
-    }
 }
 
 /**
