+++ sqlitebindings/src/nativeMain/com/birbit/sqlite3/internal/SqliteApi.kt
@@ -18,6 +18,11 @@ package com.birbit.sqlite3.internal
 import cnames.structs.sqlite3
 import cnames.structs.sqlite3_stmt
 import com.birbit.jni.jlong
+import com.birbit.sqlite3.AuthorizationParams
+import com.birbit.sqlite3.Authorizer
+import com.birbit.sqlite3.ColumnType
+import com.birbit.sqlite3.ResultCode
+import com.birbit.sqlite3.SqliteException
 import kotlin.native.concurrent.AtomicReference
 import kotlin.native.concurrent.freeze
 import kotlinx.cinterop.ByteVar
